mstpctl: move command parsing into own unit
In order to reuse the code for batch processing commands in files, move
the appropriate code in their own unit and export the needed functions
for mstpctl.
